refactor: split ItemDetailPage god file (4792 lines) into item-detail/ module
Extract JSX sections into dedicated components, shared types, utilities,
and a SegmentedControl + PreorderCloseConfirmModal component. All state
and mutations remain in the orchestrator; zero behavior change.

Old path (frontend/src/pages/admin/ItemDetailPage.tsx) now re-exports
from item-detail/index.tsx so existing imports in App.tsx are unbroken.
