+++ frontend/src/pages/admin/item-detail/index.tsx
@@ -1,0 +1,1 @@
+export { ItemDetailPage } from "./ItemDetailPage";
